ForgotPassword.test.js spurred a whole bunch of corrective development
ForgotPassword informs user if an email was sent successfully. It also
tells us if an error occurred when trying to reset the password. It
doesn't show that the email is in transit (loading icon).

LoadingScreen performs after mount check to see if we're coming from an
email. This should work in production, because the user uses the same js
client.

SignUp shows loading icon when in the process of signing up.